fix(angula): platform logic belongs to core
